Replace `Note`, `Blog`, `Article` models with single model `Post`.
IMPORTANT: Run `python manage.py migration_0062` after applying migration 0062 to create the new instances from the old models.

+++ beatonma-django/main/models/mixins/media_upload.py
@@ -27,6 +27,9 @@ class UploadedMediaMixin(models.Model):
         super().delete(*args, **kwargs)
 
     def delete_file(self):
+        if not self.file:
+            log.warning(f"No file associated with model {self}")
+            return
         if os.path.exists(self.file.path):
             os.remove(self.file.path)
             log.warning(f"Deleted file {self.file.path}")
